Desktop/FInalProject/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
PyJWT==2.10.1

# ML & Data Processing
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Astrology Calculations
pymeeus==0.5.11

# API & HTTP
httpx==0.25.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-timeout==2.2.0

# Utilities
python-json-logger==2.0.7
